_render_deploy/backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# PostgreSQL (Supabase)
psycopg2-binary==2.9.9

# Production server
gunicorn==21.2.0

# Misc
python-dotenv==1.0.0
requests==2.31.0

# SMS (optional)
twilio==8.10.0
